requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
razorpay==1.3.0
python-dotenv==1.0.0
celery==5.3.1
redis==4.6.0
PyJWT==2.8.0
bcrypt==4.0.1
gunicorn==21.2.0